+++ src/plonemeeting/portal/core/content/item.py
@@ -51,6 +51,10 @@ class IItem(model.Schema):
 
     extra_info = RichText(title=_(u"Extra info"), required=False)
 
+    plonemeeting_last_modified = schema.Datetime(
+        title=_(u"Last modification in ia.Delib"), required=True, readonly=True,
+    )
+
 
 @implementer(IItem)
 class Item(Container):