+++ src/plonemeeting/portal/core/config.py
@@ -7,3 +7,5 @@ CONTENTS_TO_CLEAN = ["Members", "events", "news"]
 
 PLONEMEETING_API_MEETINGS_VIEW = "@search_meetings"
 PLONEMEETING_API_MEETING_ITEMS_VIEW = "@search_meeting_items"
+
+API_HEADERS = {"Content-type": "application/json", "Accept": "application/json"}